requirements.txt
openai
python-dotenv
pypdf
pandas
openpyxl
PyYAML
pytest

